+++ CHANGELOG.md
@@ -8,8 +8,10 @@ BUG FIXES:
 
   - core: private boxes still referencing vagrantcloud.com will have
       their vagrant login access token properly appended
+  - core: push plugin configuration is properly validated
   - commands/push: push lookups are by user-defined name, not push
       strategy name [GH-4975]
+  - commands/push: validate the configuration
   - guests/arch: fix network configuration due to poor line breaks. [GH-4964]
   - provisioners/chef: remove Chef version check from solo.rb generation and
       make `roles_path` populate correctly